# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger",
        "title": "Gin swagger",
        "contact": {
            "name": "roy.xu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:81",
    "paths": {
        "/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "獲取用戶信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user search by id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "創建用戶",
                "parameters": [
                    {
                        "description": "create user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "a@b.com"
                },
                "username": {
                    "type": "string",
                    "example": "roy"
                }
            }
        }
    }
}